refactor: update exports and remove unused code in auth and api modules
