+++ apps/web/src/app/api/auth/[...nextauth]/route.ts
@@ -1,3 +1,3 @@
-import { handlers } from "auth";
+import { handlers } from "@repo/utils/auth";
 
 export const { GET, POST } = handlers;
