+++ packages/utils/api/index.tsx
@@ -1,3 +1,3 @@
-export * from "./index";
+export * from "./action";
 export * from "./types";
 export * from "./client-utils";
