+++ packages/utils/auth/index.tsx
@@ -1,4 +1,5 @@
+export type * from "./types.d.ts";
 export * from "./auth-actions";
-export * from "./actions";
 export * from "./auth";
 export * from "./session";
+export * from "./auth-types";
