+++ packages/utils/policies/utils.ts
@@ -1,40 +1,16 @@
 "use server";
 import { permanentRedirect, RedirectType } from "next/navigation";
-import { auth } from "../auth/auth";
 import { Policy } from "./types";
-
-export default async function PagePolicy({
-  requiredPolicies,
-  children,
-  lang,
-}: {
-  requiredPolicies: Policy[];
-  children: JSX.Element;
-  lang: string;
-}) {
-  const sessions = await auth();
-  const grantedPolicies = sessions?.grantedPolicies;
-
-  const missingPolicies = requiredPolicies.filter(
-    (policy) => !grantedPolicies?.[policy],
-  );
-  if (missingPolicies.length > 0) {
-    return permanentRedirect(`/${lang}/unauthorized`, RedirectType.replace);
-  }
-
-  return children;
-}
 
 export async function isUnauthorized({
   requiredPolicies,
   lang,
+  grantedPolicies,
 }: {
   requiredPolicies: Policy[];
   lang: string;
+  grantedPolicies?: Record<Policy, boolean> | undefined;
 }) {
-  const sessions = await auth();
-  const grantedPolicies = sessions?.grantedPolicies;
-
   const missingPolicies = requiredPolicies.filter(
     (policy) => !grantedPolicies?.[policy],
   );
